refactor(Direction): StraightDirection 패키지 이동
